Refactor start/stop button and add save button
This commit refactors the start/stop button in the Cruising Timer application and adds a save button. The start/stop button now disables the save button while the timer is running, preventing users from saving their score prematurely. When the timer is stopped, the elapsed time is calculated and displayed. If the elapsed time is less than 10 seconds, an alert is shown to prevent cheating. Otherwise, the save button is enabled, allowing users to save their score for the selected book and chapter.